LibWeb: Always draw cursor in textarea
Affinity and wrapped stops around soft line breaks not yet implemented.
