+++ Libraries/LibWeb/Painting/PaintableFragment.h
@@ -19,7 +19,7 @@ class WEB_API PaintableFragment {
     friend class PaintableWithLines;
 
 public:
-    explicit PaintableFragment(Layout::LineBoxFragment const&);
+    explicit PaintableFragment(Layout::LineBoxFragment const& fragment);
 
     Layout::Node const& layout_node() const { return m_layout_node; }
     Paintable const& paintable() const { return *m_layout_node->first_paintable(); }
@@ -42,7 +42,7 @@ public:
 
     CSSPixelRect selection_rect() const;
     CSSPixelRect range_rect(Paintable::SelectionState selection_state, size_t start_offset_in_code_units, size_t end_offset_in_code_units) const;
-
+    CSSPixelRect cursor_rect(size_t cursor_offset_in_code_units) const;
     CSSPixels width() const { return m_size.width(); }
     CSSPixels height() const { return m_size.height(); }
 
@@ -52,6 +52,7 @@ public:
 
     CSSPixels text_decoration_thickness() const { return m_text_decoration_thickness; }
     void set_text_decoration_thickness(CSSPixels thickness) { m_text_decoration_thickness = thickness; }
+    bool is_caret_anchor() const;
 
 private:
     GC::Ref<Layout::Node const> m_layout_node;
